requirements.txt
fastapi
uvicorn[standard]
python-dotenv
psycopg2-binary
rasterio
scikit-learn
google-api-python-client
google-auth
google-auth-oauthlib
pydantic
